refactor tags plugin
