+++ src/tests/smoke_login_test.ts
@@ -1,6 +1,6 @@
 import { LoginPage } from '../pages/LoginPage';
 import { allureHelper as a } from '../utils/allureHelper';
-import { tag } from '../../plugins/tagFilterPlugin';
+import { tag } from '../../plugins/tagFilter';
 
 Feature('Login page — позитивные сценарии');
